Fixed free() bug with SDL controllers
The gameLoopState was deallocated after the cleanupCallback() is called. The problem was that the struct contained sdl::Controller’s, which call SDL_GameControllerClose() in its destructor. This meant that SDL_GameControllerClose() was called after SDL was deinitialized. This commit fixes the bug by destroying the controllers before calling the cleanupCallback().