Added sfzCore Context
Added a global context to sfzCore which is explicitly set by the user before they start using the library. The primary reason for this is to make global state less implicit and give the user more control over it. This is especially important in order to support the use of sfzCore in multiple dynamic libraries at the same time.