Made ZeroG context implicit instead of explicit
I would have preferred to have the context explicit, as that would,
in theory, allow for multiple contexts at the same time. However, it
turns out to be impractical for a couple of reasons. Most prominently
we would like to overload operator new and delete because of third-party
libraries which uses standard containers and such. This is not really
possible in a good way unless we have a single global context.